refactor: events, service keys, registry

api refactor: accounts, events, registry, service keys


refactor: webhook keys


refactor: login and webhook


refactor: nodes


refactor: scale


refactor: console session


refactor: exec


refactor: update tests
