+++ controller/api/api.go
@@ -1,28 +1,20 @@
 package api
 
 import (
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"
-	"strconv"
-	"strings"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/codegangsta/negroni"
 	"github.com/gorilla/context"
 	"github.com/gorilla/mux"
 	"github.com/mailgun/oxy/forward"
-	"github.com/nu7hatch/gouuid"
-	"github.com/samalba/dockerclient"
-	"github.com/shipyard/shipyard"
 	"github.com/shipyard/shipyard/auth"
-	"github.com/shipyard/shipyard/auth/ldap"
 	"github.com/shipyard/shipyard/controller/manager"
 	"github.com/shipyard/shipyard/controller/middleware/access"
 	"github.com/shipyard/shipyard/controller/middleware/audit"
 	mAuth "github.com/shipyard/shipyard/controller/middleware/auth"
-	"github.com/shipyard/shipyard/dockerhub"
 	"github.com/shipyard/shipyard/tlsutils"
 	"golang.org/x/net/websocket"
 )
@@ -76,698 +68,6 @@ func NewApi(config ApiConfig) (*Api, error) {
 		tlsKeyPath:         config.TLSKeyPath,
 		tlsCACertPath:      config.TLSCACertPath,
 	}, nil
-}
-
-func (a *Api) addServiceKey(w http.ResponseWriter, r *http.Request) {
-	var k *auth.ServiceKey
-	if err := json.NewDecoder(r.Body).Decode(&k); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	key, err := a.manager.NewServiceKey(k.Description)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	log.Infof("created service key key=%s description=%s", key.Key, key.Description)
-	if err := json.NewEncoder(w).Encode(key); err != nil {
-		log.Error(err)
-	}
-}
-
-func (a *Api) serviceKeys(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	keys, err := a.manager.ServiceKeys()
-	if err != nil {
-		log.Error(err)
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(keys); err != nil {
-		log.Error(err)
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-}
-
-func (a *Api) removeServiceKey(w http.ResponseWriter, r *http.Request) {
-	var key *auth.ServiceKey
-	if err := json.NewDecoder(r.Body).Decode(&key); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := a.manager.RemoveServiceKey(key.Key); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	log.Infof("removed service key %s", key.Key)
-	w.WriteHeader(http.StatusNoContent)
-}
-
-func (a *Api) registries(w http.ResponseWriter, r *http.Request) {
-	registries, err := a.manager.Registries()
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := json.NewEncoder(w).Encode(registries); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) addRegistry(w http.ResponseWriter, r *http.Request) {
-	var registry *shipyard.Registry
-	if err := json.NewDecoder(r.Body).Decode(&registry); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := a.manager.AddRegistry(registry); err != nil {
-		log.Errorf("error saving registry: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	log.Infof("added registry: name=%s", registry.Name)
-	w.WriteHeader(http.StatusNoContent)
-}
-
-func (a *Api) registry(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	vars := mux.Vars(r)
-	name := vars["name"]
-
-	registry, err := a.manager.Registry(name)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := json.NewEncoder(w).Encode(registry); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) removeRegistry(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	name := vars["name"]
-
-	registry, err := a.manager.Registry(name)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := a.manager.RemoveRegistry(registry); err != nil {
-		log.Errorf("error deleting registry: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) repositories(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	vars := mux.Vars(r)
-	name := vars["name"]
-
-	if name != "" {
-		registry, err := a.manager.Registry(name)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		repos, err := registry.Repositories()
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		if err := json.NewEncoder(w).Encode(repos); err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-	}
-}
-
-func (a *Api) repository(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	vars := mux.Vars(r)
-	name := vars["name"]
-	repoName := vars["repo"]
-
-	registry, err := a.manager.Registry(name)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	repo, err := registry.Repository(repoName)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(repo); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) deleteRepository(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	name := vars["name"]
-	repoName := vars["repo"]
-
-	registry, err := a.manager.Registry(name)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := registry.DeleteRepository(repoName); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusNoContent)
-}
-
-func (a *Api) inspectRepository(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	name := vars["name"]
-	repoName := vars["repo"]
-
-	registry, err := a.manager.Registry(name)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	repo, err := registry.Repository(repoName)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(repo); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) events(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	limit := -1
-	l := r.FormValue("limit")
-	if l != "" {
-		lt, err := strconv.Atoi(l)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		limit = lt
-	}
-	events, err := a.manager.Events(limit)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(events); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) purgeEvents(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	if err := a.manager.PurgeEvents(); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	log.Info("cluster events purged")
-	w.WriteHeader(http.StatusNoContent)
-}
-
-func (a *Api) accounts(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	accounts, err := a.manager.Accounts()
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(accounts); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) saveAccount(w http.ResponseWriter, r *http.Request) {
-	var account *auth.Account
-	if err := json.NewDecoder(r.Body).Decode(&account); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := a.manager.SaveAccount(account); err != nil {
-		log.Errorf("error saving account: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	log.Debugf("updated account: name=%s", account.Username)
-	w.WriteHeader(http.StatusNoContent)
-}
-
-func (a *Api) account(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	username := vars["username"]
-
-	account, err := a.manager.Account(username)
-	if err != nil {
-		log.Errorf("error deleting account: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := json.NewEncoder(w).Encode(account); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-func (a *Api) deleteAccount(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	username := vars["username"]
-
-	account, err := a.manager.Account(username)
-	if err != nil {
-		log.Errorf("error deleting account: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := a.manager.DeleteAccount(account); err != nil {
-		log.Errorf("error deleting account: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	log.Infof("deleted account: username=%s id=%s", account.Username, account.ID)
-	w.WriteHeader(http.StatusNoContent)
-}
-
-func (a *Api) roles(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	roles, err := a.manager.Roles()
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(roles); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) role(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	vars := mux.Vars(r)
-	name := vars["name"]
-	role, err := a.manager.Role(name)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(role); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) webhookKeys(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	keys, err := a.manager.WebhookKeys()
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(keys); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) webhookKey(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	vars := mux.Vars(r)
-	id := vars["id"]
-	key, err := a.manager.WebhookKey(id)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(key); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) addWebhookKey(w http.ResponseWriter, r *http.Request) {
-	var k *dockerhub.WebhookKey
-	if err := json.NewDecoder(r.Body).Decode(&k); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	key, err := a.manager.NewWebhookKey(k.Image)
-	if err != nil {
-		log.Errorf("error generating webhook key: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	log.Infof("saved webhook key image=%s", key.Image)
-	if err := json.NewEncoder(w).Encode(key); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) deleteWebhookKey(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	id := vars["id"]
-	if err := a.manager.DeleteWebhookKey(id); err != nil {
-		log.Errorf("error deleting webhook key: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	log.Infof("removed webhook key id=%s", id)
-	w.WriteHeader(http.StatusNoContent)
-}
-
-func (a *Api) login(w http.ResponseWriter, r *http.Request) {
-	var creds *Credentials
-	if err := json.NewDecoder(r.Body).Decode(&creds); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	loginSuccessful, err := a.manager.Authenticate(creds.Username, creds.Password)
-	if err != nil {
-		log.Errorf("error during login for %s from %s: %s", creds.Username, r.RemoteAddr, err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if !loginSuccessful {
-		log.Warnf("invalid login for %s from %s", creds.Username, r.RemoteAddr)
-		http.Error(w, "invalid username/password", http.StatusForbidden)
-		return
-	}
-
-	// check for ldap and autocreate for users
-	if a.manager.GetAuthenticator().Name() == "ldap" {
-		if a.manager.GetAuthenticator().(*ldap.LdapAuthenticator).AutocreateUsers {
-			defaultAccessLevel := a.manager.GetAuthenticator().(*ldap.LdapAuthenticator).DefaultAccessLevel
-			log.Debug("ldap: checking for existing user account and creating if necessary")
-			// give default users readonly access to containers
-			acct := &auth.Account{
-				Username: creds.Username,
-				Roles:    []string{defaultAccessLevel},
-			}
-
-			// check for existing account
-			if _, err := a.manager.Account(creds.Username); err != nil {
-				if err == manager.ErrAccountDoesNotExist {
-					log.Debugf("autocreating user for ldap: username=%s access=%s", creds.Username, defaultAccessLevel)
-					if err := a.manager.SaveAccount(acct); err != nil {
-						log.Errorf("error autocreating ldap user %s: %s", creds.Username, err)
-						http.Error(w, err.Error(), http.StatusInternalServerError)
-						return
-					}
-				} else {
-					log.Errorf("error checking user for autocreate: %s", err)
-					http.Error(w, err.Error(), http.StatusInternalServerError)
-					return
-				}
-			}
-
-		}
-	}
-
-	// return token
-	token, err := a.manager.NewAuthToken(creds.Username, r.UserAgent())
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(token); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) changePassword(w http.ResponseWriter, r *http.Request) {
-	session, _ := a.manager.Store().Get(r, a.manager.StoreKey())
-	var creds *Credentials
-	if err := json.NewDecoder(r.Body).Decode(&creds); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	username := session.Values["username"].(string)
-	if username == "" {
-		http.Error(w, "unauthorized", http.StatusInternalServerError)
-		return
-	}
-	if err := a.manager.ChangePassword(username, creds.Password); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) hubWebhook(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	id := vars["id"]
-	key, err := a.manager.WebhookKey(id)
-	if err != nil {
-		log.Errorf("invalid webook key: id=%s from %s", id, r.RemoteAddr)
-		http.Error(w, err.Error(), http.StatusNotFound)
-		return
-	}
-	var webhook *dockerhub.Webhook
-	if err := json.NewDecoder(r.Body).Decode(&webhook); err != nil {
-		log.Errorf("error parsing webhook: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if strings.Index(webhook.Repository.RepoName, key.Image) == -1 {
-		log.Errorf("webhook key image does not match: repo=%s image=%s", webhook.Repository.RepoName, key.Image)
-		http.Error(w, "not found", http.StatusNotFound)
-		return
-	}
-	log.Infof("received webhook notification for %s", webhook.Repository.RepoName)
-	// TODO @ehazlett - redeploy containers
-}
-
-func (a *Api) nodes(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	nodes, err := a.manager.Nodes()
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(nodes); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) node(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	vars := mux.Vars(r)
-	name := vars["name"]
-	node, err := a.manager.Node(name)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(node); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) scaleContainer(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	vars := mux.Vars(r)
-	containerId := vars["id"]
-	n := r.URL.Query()["n"]
-
-	if len(n) == 0 {
-		http.Error(w, "you must enter a number of instances (param: n)", http.StatusBadRequest)
-		return
-	}
-
-	numInstances, err := strconv.Atoi(n[0])
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	if numInstances <= 0 {
-		http.Error(w, "you must enter a positive value", http.StatusBadRequest)
-		return
-	}
-
-	result := a.manager.ScaleContainer(containerId, numInstances)
-	// If we received any errors, continue to write result to the writer, but return a 500
-	if len(result.Errors) > 0 {
-		w.WriteHeader(http.StatusInternalServerError)
-	}
-	if err := json.NewEncoder(w).Encode(result); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-}
-
-func (a *Api) createConsoleSession(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	vars := mux.Vars(r)
-	containerId := vars["container"]
-	// generate token
-	u4, err := uuid.NewV4()
-	if err != nil {
-		log.Errorf("error generating console session token: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	token := u4.String()
-
-	cs := &shipyard.ConsoleSession{
-		ContainerID: containerId,
-		Token:       token,
-	}
-
-	if err := a.manager.CreateConsoleSession(cs); err != nil {
-		log.Errorf("error creating console session: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := json.NewEncoder(w).Encode(cs); err != nil {
-		log.Errorf("error encoding console session: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	log.Debugf("created console session: container=%s", cs.ContainerID)
-}
-
-func (a *Api) consoleSession(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("content-type", "application/json")
-
-	vars := mux.Vars(r)
-	token := vars["token"]
-
-	cs, err := a.manager.ConsoleSession(token)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := json.NewEncoder(w).Encode(cs); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) removeConsoleSession(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	token := vars["token"]
-
-	cs, err := a.manager.ConsoleSession(token)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	if err := a.manager.RemoveConsoleSession(cs); err != nil {
-		log.Errorf("error removing console session: %s", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (a *Api) execContainer(ws *websocket.Conn) {
-	qry := ws.Request().URL.Query()
-	containerId := qry.Get("id")
-	command := qry.Get("cmd")
-	ttyWidth := qry.Get("w")
-	ttyHeight := qry.Get("h")
-	token := qry.Get("token")
-	cmd := strings.Split(command, ",")
-
-	if !a.manager.ValidateConsoleSessionToken(containerId, token) {
-		ws.Write([]byte("unauthorized"))
-		ws.Close()
-		return
-	}
-
-	log.Debugf("starting exec session: container=%s cmd=%s", containerId, command)
-	clientUrl := a.manager.DockerClient().URL
-
-	execConfig := &dockerclient.ExecConfig{
-		AttachStdin:  true,
-		AttachStdout: true,
-		AttachStderr: true,
-		Tty:          true,
-		Cmd:          cmd,
-		Container:    containerId,
-		Detach:       true,
-	}
-
-	execId, err := a.manager.DockerClient().ExecCreate(execConfig)
-	if err != nil {
-		log.Errorf("error calling exec: %s", err)
-		return
-	}
-
-	if err := a.hijack(clientUrl.Host, "POST", "/exec/"+execId+"/start", true, ws, ws, ws, nil, nil); err != nil {
-		log.Errorf("error during hijack: %s", err)
-		return
-	}
-
-	// resize
-	w, err := strconv.Atoi(ttyWidth)
-	if err != nil {
-		log.Error(err)
-		return
-	}
-
-	h, err := strconv.Atoi(ttyHeight)
-	if err != nil {
-		log.Error(err)
-		return
-	}
-
-	if err := a.manager.DockerClient().ExecResize(execId, w, h); err != nil {
-		log.Errorf("error resizing exec tty: %s", err)
-		return
-	}
-
 }
 
 func (a *Api) Run() error {
